nahodeny dialog na vymazanie dispecera

+++ desktop/mantraDesktop/src/app/components/dispecer/the-dispecer/the-dispecer.component.ts
@@ -3,6 +3,9 @@ import Dispecer from "../../../models/Dispecer";
 import {DispecerService} from "../../../services/dispecer.service";
 import {FormBuilder} from "@angular/forms";
 import {DataService} from "../../../data/data.service";
+import {DeleteCarDialogComponent} from "../../dialogs/delete-car-dialog/delete-car-dialog.component";
+import {MatDialog} from "@angular/material/dialog";
+import {DeleteDispecerComponent} from "../../dialogs/delete-dispecer/delete-dispecer.component";
 
 @Component({
   selector: 'app-the-dispecer',
@@ -14,14 +17,24 @@ export class TheDispecerComponent implements OnInit {
   @Input() dispecer: Dispecer;
   dispecerovForm: any;
 
-  constructor(private dispecerService: DispecerService, private field: FormBuilder, public dataService: DataService) { }
+  constructor(private dispecerService: DispecerService, private field: FormBuilder, public dataService: DataService,
+     private dialog: MatDialog) { }
 
   ngOnInit(): void {
 
   }
 
   deleteDispecer(){
-    this.dispecerService.deleteDispecer(this.dispecer.id);
+    const dialogRef = this.dialog.open(DeleteDispecerComponent, {
+      data: {dispecer: this.dispecer}
+    });
+    dialogRef.afterClosed().subscribe(value => {
+      if (value === undefined){
+        return;
+      }else {
+        this.dispecerService.deleteDispecer(this.dispecer.id);
+      }
+    });
   }
 
 }
